platform/requirements.txt
# Core web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
PyYAML>=6.0
pytz>=2023.3

# Market data (Alpaca)
alpaca-py>=0.13.0
alpaca-trade-api>=3.1.0
websockets==10.4  # Must stay 10.x for alpaca-trade-api compatibility

# ML / Kronos model (optional - platform works without these)
torch>=2.0.0
transformers>=4.30.0

# LLM analysis (Gemini)
google-generativeai>=0.3.0

# Twitter/X API client
httpx>=0.27.0

# Database (async PostgreSQL + pgvector)
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
pgvector>=0.3.0
alembic>=1.13.0

# Embedding model (for vector memory)
sentence-transformers>=3.0.0